requirements.txt
torch>=2.0
torchvision>=0.15
timm>=0.9
scikit-learn>=1.2
matplotlib>=3.7
pillow>=9.5
numpy>=1.24
pandas>=2.0
tqdm>=4.65
ipywidgets>=8.0
nbformat>=5.8
